refactor: move Explain state to its own file
