+++ src/states/game.ts
@@ -3,12 +3,13 @@ import Phaser from 'phaser';
 import { BASE_SIZE } from '../variables';
 import { Victoire } from './victoire';
 import { Load } from './load';
-import { Explain, Menu } from './menu';
+import { Menu } from './menu';
 import { Edit } from './edit';
 import { Scores } from './scores';
 import { Play } from './play';
 import { Boot } from './boot';
 import { Langue } from './langue';
+import { Explain } from './explain';
 
 // Variable de jeu
 export const game = new Phaser.Game(BASE_SIZE, BASE_SIZE, Phaser.CANVAS, 'cycleCanvas');
