+++ src/states/menu.ts
@@ -1,48 +1,9 @@
 // On crée l'état du menu
-import { cliquable, difficultes, reset } from '../functions';
+import { cliquable, reset } from '../functions';
 import { BASE_SIZE, platformSizeInPx, Stats } from '../variables';
 import { lang } from '../i18n';
 import { lang as selectedLang } from './langue';
-import Phaser from 'phaser';
 import { LOGO_C, LOGO_C_MASK, LOGO_WITHOUT_C } from '../resoucesNames';
-
-export const Explain = {
-    create: function() {
-        // Un équivalent d'écouteur d'évènement (si on veut)
-        this.space_key = this.game.input.keyboard.addKey(Phaser.Keyboard.SPACEBAR);
-
-        // Retour au menu avec Échap
-        this.esc_key = this.game.input.keyboard.addKey(Phaser.Keyboard.ESC);
-
-        // Présentation du jeu
-        cliquable(BASE_SIZE / 2, BASE_SIZE / 2.5, lang[selectedLang].Presentation, 25, 0.5, 0.5, 0, 500, null, this);
-
-        // Difficulté
-        this.choix_diff = [];
-        difficultes(this, null);
-
-        if (isNaN(Stats.difficulty)) Stats.difficulty = 1;
-
-        this.choix_diff[Stats.difficulty].setShadow(0, 0, 'rgba(0, 0, 0, 1)', 5);
-    },
-    souligne: function(sprite) {
-        for (var val in this.choix_diff) {
-            this.choix_diff[val].setShadow(0, 0, 'rgba(0, 0, 0, 0)', 5);
-        }
-
-        sprite.setShadow(0, 0, 'rgba(0, 0, 0, 1)', 5);
-
-        reset(Stats);
-        Stats.difficulty = sprite.difficulty;
-    },
-    update: function() {
-        // Retour au menu
-        if (this.esc_key.isDown) this.game.state.start('Menu');
-
-        // Démarrage du jeu
-        if (this.space_key.isDown) this.game.state.start('Play');
-    },
-};
 
 // Précédant les explications : choix entre editor et game
 export const Menu = {
